Make deployment behaviour configurable [git or dir]
We'd like this as the `git` style copies across the whole git archive,
needed when a project has dependencies that aren't children of it in the
directory hierarchy. However, we'd like to retain the `dir` style still
that only copies across the local directory and its descendants as that
may be far more efficient when a project can be built this way.