Totally change how "dp deploy" works...
Before this commit, dp created a deployment tarball from the *current
working directory*, irrespective of what branch you told it to deploy,
and irrespective of any uncommitted or even unstaged changes in the
working directory.  This commit attempts to make the behaviour of "dp
deploy" a little more sane.

The new "dp deploy" does the following:

- Clones the Git repository where "dp deploy" is run on the remote
  machine.  The clone appears in the "versions/<date>" directory for the
  application being deployed.  (The remote machine thus needs to have
  the appropriate SSH keys in place to allow this to work.)

- Sets up Git submodules and checks out the requested branch.

- Uses the ".dp/compile" script to build the application.

- Creates symbolic links as required.

This procedure means that the deployments ALWAYS happen from a
well-defined branch in the Git repository.