VM builds
If the `DP_VM_BUILDS` environment variable is set to `true`, builds for
`dp deploy` will be performed within a Vagrant virtual machine instead
of on the target machine.  Build artefacts to generate the upload
tarball will be collected using an `extract` script that should be in
the `.dp` directory alongside the `compile` script.